+++ iam/organization/application/create/create.command.ts
@@ -1,7 +1,6 @@
-import { ICommand } from '@nestjs/cqrs';
 import { Identifier } from 'src/shared/core/domain/identifier';
 
-export class CreateOrganizationCommand implements ICommand {
+export class CreateOrganizationCommand {
   constructor(
     public readonly organizationId: Identifier,
     public readonly name: string,
