+++ iam/organization/application/delete/delete.command.ts
@@ -1,0 +1,3 @@
+export class DeleteOrganizationCommand {
+  constructor(public readonly organizationId: string) {}
+}
