+++ iam/organization/application/update/update.use-case.ts
@@ -1,21 +1,22 @@
 import { Inject, Injectable } from '@nestjs/common';
+import { UpdateOrganizationCommand } from './update.command';
 import { ORGANIZATION_STORE, OrganizationStore } from '../../domain/organization.store';
-import { DeleteOrganizationCommand } from './delete-organization.command';
 import { CustomException } from 'src/shared/exception/custom-exception';
 import { CustomExceptionCode } from 'src/shared/exception/custom-exception-code';
 
 @Injectable()
-export class DeleteOrganizationUseCase {
+export class UpdateOrganizationUseCase {
   constructor(
     @Inject(ORGANIZATION_STORE)
     private readonly organizationStore: OrganizationStore,
   ) {}
 
-  async execute(command: DeleteOrganizationCommand): Promise<void> {
+  async execute(command: UpdateOrganizationCommand): Promise<void> {
     const organization = await this.organizationStore.loadById(command.organizationId);
     if (!organization) throw new CustomException(CustomExceptionCode.ORGANIZATION_NOT_FOUND);
 
-    const { organizationId } = command;
-    await this.organizationStore.deleteById(organizationId);
+    organization.update(command.name, command.contact);
+
+    await this.organizationStore.updateById(organization);
   }
 }
