+++ iam/organization/domain/organization.view.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 
-export class OrganizationModel {
+export class OrganizationView {
   @ApiProperty()
   name: string;
