+++ iam/organization/presentation/organization.controller.ts
@@ -1,20 +1,23 @@
 import { Body, Controller, Patch, UseGuards } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
-import { UpdateOrganizationUseCase } from '../application/update-organization/update-organization.use-case';
-import { UpdateOrganizationDto } from './dto/request/update-organization.request.dto';
 import { AuthGuard } from '@nestjs/passport';
-import { OrganizationCommandDocs } from './organization.command.docs';
+import { OrganizationDocs } from './organization.docs';
 import { Organization, OrganizationPayload } from 'src/shared/core/presentation/organization.decorator';
+import { UpdateOrganizationUseCase } from '../application/update/update.use-case';
+import { UpdateOrganizationReqDto } from './dto/request/update-organization.request.dto';
 
 @ApiTags('organization')
-@Controller('organization')
-export class OrganizationCommandController {
+@Controller('organization-temp')
+export class OrganizationController {
   constructor(private readonly updateOrganizationUseCase: UpdateOrganizationUseCase) {}
 
   @Patch()
   @UseGuards(AuthGuard('jwt-access'))
-  @OrganizationCommandDocs('update')
-  async update(@Organization() organization: OrganizationPayload, @Body() dto: UpdateOrganizationDto): Promise<void> {
+  @OrganizationDocs('update')
+  async updateOrganization(
+    @Organization() organization: OrganizationPayload,
+    @Body() dto: UpdateOrganizationReqDto,
+  ): Promise<void> {
     await this.updateOrganizationUseCase.execute({
       organizationId: organization.organizationId,
       name: dto.name,
