--- src/organization/command/application/delete-organization/delete-organization.command.ts
@@ -1,5 +1,0 @@
-import { ICommand } from '@nestjs/cqrs';
-
-export class DeleteOrganizationCommand implements ICommand {
-  constructor(public readonly organizationId: string) {}
-}
